AtmService для любого количества и вида купюр.

+++ hw07-atm-emulator/src/main/java/org/example/atm/service/AtmService.java
@@ -8,6 +8,6 @@ public interface AtmService {
 
     int getTotalAmount();
 
-    CashBundle getCash(int amount);
+    CashBundle getCash(long amount);
 
 }
